fix(gh): specify config options for actions/stale
This changes the time of closing from 60 days to 30 days. It also lets us apply a "no-stale" label or a milestone to prevent an issue or pull request from going stale. Draft pull requests will also not go stale (but this may change if we do get any of those!)

Most of these are default configuration options, but as this project grows I will likely change these to be more specific. This also prevents a default configuration change from negatively impacting the repository.

Signed-off-by: JBMagination <jbmagination@jbmagination.com>